backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
python-multipart>=0.0.9
python-dotenv>=1.0.0,<2.0.0
pandas>=2.2.0,<3.0.0
pydantic>=2.9.0,<3.0.0
pydantic-settings>=2.6.0,<3.0.0
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.3.0
chromadb>=0.5.0,<0.6.0
sentence-transformers>=3.0.0,<4.0.0
openai>=1.50.0,<2.0.0
